sh1106.py: removing calls to Pin.high and Pin.low and support for WiPy 2 et. al.
a) The methods Pin.high() and Pin.low() were abandoned from the micropython.org
port of MicroPython. Therefore, the basic pin(1) and pin(0) versions
had to be used.
b) The pycom.io port of Micropython does not support the i2c.start() and
i2c.write() methods. Therefore alternatively i2c.writeto() had to be used.
Thanks to @pacmac for implementing and testing it.