Allow rotating by 90 degrees
Since the SH1106 has no hardware flag for doing this, we're doing it in
software by creating a second buffer. One will be the user-facing one
that receives all the render commands (`self.renderbuf`), while the
other one will be constructed by remapping the bytes as required during
`show()` (`self.displaybuf`). The second buffer will only be created
when 90° rotation is enabled, because it essentially doubles the RAM
required.

This renames the `rotate()` method to `flip()` in order to (hopefully)
be less confusing. 90° rotation has to be enabled when constructing the
instance, it cannot be toggled afterwards, because of the additional
buffer required. For backwards compatibility, `rotate()` is still
available as an alias to `flip()`.

The code has only been tested with a 128×64 display, because that's
what's been available to me.