sh1106.py: Use a bit fied for pages-to-update.
This is a slight variant of the optimization implemented by @CRImier,
using a bit fiels instead of a list to memoroze the pages, that have
to be updated. That avoids memory allocations. The code size is a few
bytes less, the execution times are the same.

The partial update introduced by @CRImier is a huge improvement for
small changes to the screen when using slow controllers like the
ESP8266. A full screen update take ~80ms, a single page takes about
10ms. So at best the speed increase is a factor of 8. The downside is
an increase in code size of about 1k, which again may be an issue
with the ESP8266. If that is a limiting factor, frozen bytecode
can be used.