Update sh1106.py
This pull request/ change is intended to fix issue #28.
I have tested this code in the [MicroPython SH1107 driver](https://github.com/peter-l5/SH1107) which is based on this one, but not actually with this driver itself directly as I do not have an SH1106 display. However as it relates to the interface with the FrameBuffer class, rather than the SH1106, I think it should work without any problems.